[refactor] : 1. PostRepository -> JpaRepository 인터페이스로 변경 2. BaseEntity 추가하여 Post Entity 상속
