+++ src/main/java/com/study/spring/controller/MemberController.java
@@ -2,8 +2,8 @@ package com.study.spring.controller;
 
 import com.study.spring.dto.AddUserRequest;
 import com.study.spring.global.response.ApiResponse;
-import com.study.spring.global.response.CustomException;
-import com.study.spring.global.response.ErrorCode;
+import com.study.spring.global.Exception.CustomException;
+import com.study.spring.global.Exception.ErrorCode;
 import com.study.spring.service.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/member")
+@RequestMapping("/api")
 public class MemberController {
 
     private final MemberService memberService;
 
-    @PostMapping("/signup")
+    @PostMapping("/member/signup")
     public ApiResponse<?> create(@RequestBody @Valid AddUserRequest addUserRequest, BindingResult bindingResult) {
 
         if (bindingResult.hasErrors()) {
