+++ src/main/java/com/study/spring/domain/Post.java
@@ -1,25 +1,21 @@
 package com.study.spring.domain;
 
 import com.study.spring.dto.req.PostRequestDTO;
+import com.study.spring.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import org.springframework.security.crypto.password.PasswordEncoder;
-
-import java.time.LocalDateTime;
 
 
 @Entity
 @Getter
 @Setter 
 @Builder
-@NoArgsConstructor
-@AllArgsConstructor
+@NoArgsConstructor  // 기본 생성자 자동 생성 
+@AllArgsConstructor // 모든 필드를 매개변수로 받는 생성자 자동 생성
+// @RequiredArgsConstructor -> 필수 인자를 받는 생성자를 자동으로 생성 final 필드 또는 @NonNUll 붙은 필드만 생성
 // @DynamicInsert  // entity를 save시 null 값은 배제하고 insert 실행
-@EntityListeners(AuditingEntityListener.class)  // 자동 날짜 생성
-public class Post {
+public class Post extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "post_id")
@@ -30,11 +26,6 @@ public class Post {
     private String content;
     private String password;
     private String author;
-
-    @CreatedDate
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
 
     // columnDefinition 제약사항만 넣어줄뿐, null 값을 defalut value로 치환해주지는 않음
     @Column(name = "POST_STATUS")
