+++ src/main/java/com/study/spring/global/entity/BaseEntity.java
@@ -1,0 +1,22 @@
+package com.study.spring.global.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass   // JPA 엔티티의 부모 클래스임 선언 및 직접 테이블로 매핑되지 않은 클래스
+@EntityListeners(AuditingEntityListener.class)  // 자동 날짜 생성
+public abstract class BaseEntity {
+
+    @CreatedDate
+    private LocalDateTime createdAt;
+    @LastModifiedDate
+    private LocalDateTime updatedAt;
+}
